New version of UARTP1ReaderComponent
Timestamp and GasTimestamp now also included as TextSensor. New version of the component allows for more user configurability as well. Unwantend properties can be set to NULL/nullptr to be discarded. This requires a bit more configuration in the .yaml files (compare between commits to see what I mean) but increases customizability for users. New component has been applied to all the different configurations. Updated the image in the wiki to reflect the increase in available MQTT topics.